split EditorObject.cs into multiple files for the different classes
